Implementation and use of preferences finished (the notification and mediaplayer have changed the behaviour), distance function fixed (now it works as it should be and it gives you the correct location), implemented logic of visualize the data in the recycleview only when you are watching it and when you resume the activity, new image and fixed the view of the notification and more...

+++ app/src/main/java/com/android4dev/CityTourApp/models/touristicPlace.kt
@@ -14,5 +14,5 @@ class TouristicPlace (
     var score: Double,
     var audioGuideUrl: String,
     val type: TP_Type,
-    var distance: Double? = null
+    var distance: Float? = null
 ) : Serializable
